Added group/selection email and sms sending.
Added options for user to send emails and sms texts to
all members of a group (or selection) at once. If there
are multiple emails/phones for a single contact, a picker
is displayed to specify which emails/phones to use.

Test: Verified all features are working as intended.
Menu options are only visible when relevant and edge
case of some or all users not having emails/phones
are handled by showing a toast to the user.

Bug: 31648014
Change-Id: I38066cf3be57bf205f7a3721d0064bb716e8a43f