Prevent multiple editors from opening
Before we start the editor, destroy the loader so it isn't listening
for changes to CP2 which would cause it to reload and go through the
loadEditor code path again.

Test:
Added a manual forceLoad (after the loadEditor call) in
onLoadFinished and saw that it started more than one editor.
Added the destroyLoader call and confirmed even with the above only
one instance opened.

Bug: 32119921
Change-Id: Ib683cb60175fd816b38bd8b6d12b445ca236cd9e