Item multi select now has defaults checked.
Added ability to send a list of ids to a
multi select picker to precheck defaults.
Defaults will preselect if the aggregate
contact has a single email/phone, or if
there is a super primary set for the contact.

Test: Verified all features are working as intended.
Picker now shows all contact methods for group/selection,
which fixes some strange selection behavior.

Bug: 31648014
Change-Id: I21ee4e79f8c1d08d41b4edcdef62aa5b2f2eb4b9