Always write current state in EditorFragment
When the choose default account activity dialog is shown,
mState in EditorFragment is empty, so it doesn't get written to the
out bundle. When we come back to the fragment save state exists
and null is read.

Test:
Followed repro steps in bug. Only repro'd when don't keep activities
was enabled.

Bug: 32118572
Change-Id: I80e972cf18ac1b460a849a8e5a2cc8853bac62c4