Forward intent extras to editor activity from springboard
Apply the extras once we have selected a raw contact to edit.

Test:
Selected an existing contact to add to and saw passed in
data was correctly applied.
Tested cases where editor auto loads, and picker dialog shows up.
In the case of the picker made sure that selecting any of them
would have the data passed through.

Bug: 32345528
Change-Id: I9a4088b3358ab532362c878394c657781965fb03