Change aggregation suggestions to be raw contacts
Each suggestion is a raw contact that belongs to the same account
of the currently edited raw contact.
Surfaced data is the raw contact's data.
Pass the current changes to the newly selected raw contact.
Remove a few unneeded methods.
Removed the join option for tapping on a suggestion since all raw
contacts selected from here will be editable.

Since QuickContact depends on the aggregation suggestion engine to
return contact level suggestions also removed everything related to
it there since it's currently disabled.
b/29216287 tracks plans bring it back but it probably won't be in the
form of using the suggestion engine to populate the card, so it felt
okay to remove the code. Layouts were left in case they can be reused
 later.

Test:
Opened QuickContact
Tested the suggestions for a contact with multiple raw contacts both
in and out of the current account and made sure only ones in the
current accountwere shown.
Checked contact with raw contacts which have different names, made
sure those displayed correctly. CP2 doesn't store the alternative
versions of raw contact names so the primary one is always shown
which is a non-regression.

Bug: 32277025
Bug: 32278866
Bug: 31088704
Bug: 31704115
Change-Id: Ibd4bdaaf100ffd5524eebac300d23763bbe6a22d