Changed how defaults are picked for group send.
In the email/phone picker there will always be a defaut
selected for contacts. Defaults are based on if the
contact has a primary set, if the contact only has a
single email/phone, or lastly, which of the emails/
phones has the highest contact frequency.

Test: Manual test, defaults are now selected by picking
the item with the highest contact frequency.

Bug: 32460061
Change-Id: I1c0667d1a9825499cf6538df63270f2d0dfb21bf