Ignore read-only raw contacts in picker modal pt. 1
Drop read-only contacts unless the SpringboardActivity is configured
not to.
Changed to using AsyncTaskLoader that returns a RawContactsMetadata
object which has all the information needed by the picker dialog.

Ignore legacy edit URIs and send silent feedback if they are given.

Test:
Manually checked the following scenarios has the correct behavior:
  * Single writable raw contact
  * Multiple writable raw contacts
  * 1 read-only 1 writable
  * 1 read-only
  * 2 read-only
  * Test app usage of legacy URI

Bug: 32640214
Bug: 32240763
Change-Id: Ia3542b927ea615a2b2d07f86cdc1ff8c91f99ba6