Use toolbar in editor
Lets us change content description of the navigation icon a bit
easier.
When it's an X read out "Cancel button".
When it's a back arrow read out "Back button".
Left the toolbar within the activity instead of putting it in the
fragment to reduce risk/testing needed for making this change.

Test: Manually verified for the above for:
  * editing a contact
  * adding a new contact
  * viewing a read-only contact

Bug: 32950789

Change-Id: I125728786bdf3872e141863f60d38b34e01bb757