Add null check to Uri when getting the name to toast
When all contact details are removed from a contact when there are
no accounts (potentially if it's the device account) the contact is
fully deleted, so the Uri is null when we try to get the name to
display. Bail early and return null in this case.

Test: Verified no crash and consistent behavior with previous releases
when removing all contact info from a contact when no accounts are
present and when deleting all info from the ME profile with Google+
contacts sync turned off.

Bug: 32982336
Change-Id: I0fd236080750a5040a5c9ffd80a48a37cb1f8ae3