Ignore display name when creating aggregation uris
When all fields of a name are removed, the display name is used
to find an aggregation suggestions. Since we removed support for
editing the display name it will always be the initial value, causing
a lookup to be done for the initial name if all characters are
removed.

Test: Verified removing the name doesn't trigger a suggestion for
the initial display name.

Bug: 33050912
Change-Id: Id45ca55a3fa522048070d37965f512b1d04c12a3