Fix rotation crash in springboard
Save mResult in the instance bundle. It used to always be delivered
from the loader but it's possible for the springboard to not be the
top activity now so it can get recreated without getting a result
from the loader.

Test: Manually verified rotating the screen in the "choose a contact
to link" activity doesn't crash the app and the link happens as
expected.

Bug: 33196170
Change-Id: I0a5b21aed625059a1f65dc2e26de85af3feb5ebe