Fix editor crash when rotating during contact insert
Remember the current account being saved to when setting state
for a new contact. Otherwise when during an insert it will never
get set unless you select a new account from the switcher dropdown.
Since it's null once the account load is finished it goes through
the set state for a new contact code path again which adds another
RawContactDelta to the list which breaks some of our assumptions
regarding the size of mState.

Test: Manually verified no crash occurs after rotation when adding
a new contact from:
  * the fab
  * recommendations assistant
  * third party apps (dialer/messenger)

Bug: 34126254
Bug: 34131358
Change-Id: I874cc592f94584a98e54010751c557cfd3b0d389