Fix gradle release build (1/2)
Jack toolchain in build tools 26.0.0 has some issues in obfuscation that
cases runtime errors in guava collect and concurrent. Add proguard flags
to workaround this. Note that we are still getting a dramatic
improvement in apk size since release build in jack is 7.1 mbs but old
make builds was 11 mbs.

Test:
  Run installGoogleRelease and sanity checks
  Built GoogleContacts user in make
  Ran make GoogleContactsTests
  Ran gradle connectedGoogleDebugAndroidTest

Bug:34127945
Bug:34917059
Change-Id: Ifebe6907cae87ac574bae1f7aefd83a79258badb