Photo view in editor fixes
Use the height/width returned from getHeight and getWidth to
resize the photo view. Since it takes up all available space
initially these two values essentially function as the content
view height/width.

Test: Manually verified photo view sizing on M, N, and O with
rotation.
Also verified editor alignment in multiwindow mode on N.
One caveat is when rotating to landscape with the keyboard up on
O only, there's still a little bit of space at the bottom of the
photo view. This doesn't repro when loading initially in landscape
or rotating into landscape without the keyboard up. So the bug
isn't fixed entirely but it handles the more common cases.

Bug: 33922016
Bug: 27887505
Change-Id: I89ad06bfb07d8891c6ecb57cc7cb0452a2112385