Adopt O shortcut changes for Intent.ACTION_CREATE_SHORTCUT
If we're on O, use ShortcutManager's .createShortcutResultIntent
and return that to the launcher.
Move ShortcutIntentBuilder out of list package since it doesn't
really make sense in there. Also allows us to keep DynamicShortcuts'
methods package private.

Test: Manually verify shortcuts created by long pressing on the
home screen (Contacts 1x1 widgets) are the newer app shortcuts on O
and are legacy shortcuts on pre-O (L, M, and N).

Bug: 36032908
Change-Id: Ic78c21daf223b59b45cbc98ceea2726fc29c055c