Cannot open date before 1970/01/01 from quickcontact screen
The pre-install calendar support to jump the oldest date is 1970/01/01.
If date set before 1970/01/01, it can't jump to the date.

When we save a birthday or anniversary event for a contact, we
set the year to 0, this means when a user for example clicks a birthday
we will open the day of the coming birthday.

Bug: TBD
Test: Manual
Change-Id: I68d207cddbab53b97c1368048a50cda5d6bd5aa2