Add alternative video call to QC action
Force the clickable icons in the entry card to be 48dp.
Add the reachability DataItem to the PhoneDataItem so we
have all the necessary data for creating an Entry.
Add a boolean for whether we should color the "third icon", we don't
want to tint it when using the icon from a third party app.
Since we're adding the reachable action to the phone number row,
skip creating the custom row.

Consider cleaning up the Entry constructor (maybe in 2.1), 5 of the 6
uses pass in a lot of null/default values. Possibly use a Builder
or removing the finals from the fields.

Test: Manually verified phone numbers which are reachable on Duo, have
the corresponding action and the custom row isn't visible.
https://screenshot.googleplex.com/vqoCjGwCpYd

Bug: 36874501
Change-Id: Ie3d0fb6ca6491e0fd3d890d47980cc845b9241db