Backport cl/159625495
*** Original change description ***

Include the phone number as part of the shortcut id when creating dial
or sms shortcuts

The id was previously just the type of shortcut it was (call or dial)
concatenated with the lookupKey for the contact. So when creating two
call or two sms shortcuts for the same contact with different numbers,
the number for the first one created will always be used since the
shortcut manager would see the same ID already existed and so an
identical shortcut would be placed.
This CL includes the hashCode for the phone URI in the shortcut ID.

Tested:
Manually verified creating dial and sms shortcuts for a contact with two
numbers actually creates two different shortcuts that will call/message
the different numbers.

***

Test: Same verification as the original change.
Bug: 62786173
Change-Id: I9b9ad71401f44dc8e84acde92263d25d0b3ef5c7