Fixed to Added the function to add the Phonetic name into Phonetic field automaticaly.
This fix is to add the Phonetic name automatically to Phonetic name field when user inputs a Name in Japanese language mode.

Enter in the name field and automatically enter to Phonetic field when the input charactor is confirmed.
The cases that will be entered to the Phonetic field are as follows.
1. Select enter on the keyboard
2. Select for prediction conversion candidate
3. Tap outside the input field.

Change-Id: I6a04a971814a1e84cd2db7acc5d7c80c1ca79074
Signed-off-by: yoichi kakimoto <youichi.kakimoto.gt@kyocera.jp>