Add android.test.mock/stubs dependency
In preparation for the removal of the android.test.mock classes from
the android.jar this adds dependencies on android.test.mock/stubs to
ensure this code will continue to compile.

The following change descriptions were generated automatically and so
may be a little repetitive. They are provided to give the reviewer
enough information to check the comments match what has actually been
changed and check the reasoning behind the changes.

* tests/Android.mk
    Added 'android.test.mock.stubs' to LOCAL_JAVA_LIBRARIES because
    ContactsTests's will need them in order to compile once its classes
    are removed from the current SDK on which it currently depends.

Bug: 30188076
Test: make dist
Change-Id: I3944e6f5049c2c369c5e18c396208dea8c255b83