Send broadcast to request high res photo sync
For GoogleAccountTypes don't start the service to notify the
adapter that the contact was viewed. Instead send a broadcast
to request a high res photo sync.
Fallback to original code path if there are no receivers.

Test: Manual test with debug logging enabled. Viewing a low res
photo synced the high res photo after the broadcast was received
and the job intent service was run.
Tested on Q build: QP1A.190430.002

Bug:131899328
Change-Id: I85212d6173567a305868a631a610f003b8cbc7a7
Merged-In: I85212d6173567a305868a631a610f003b8cbc7a7