Add required flag to registerReceiver call in AccountTypeManager
Android T adds support to allow a runtime receiver to be registered as
not exported, but to ensure apps can take advantage of this, calls to
registerReceiver must specify a flag indicating whether the receiver
should be exported for apps targeting T+ that are registering for
unprotected broadcasts. This commit adds the RECEIVER_EXPORTED flag
to the call to registerReceiver in AccountTypeManager when
registering for the ACTION_SIM_ACCOUNTS_CHANGED action
as this is broadcast from the ContactsProvider.

Bug: 161145287
Test: Build
Change-Id: I713c096b7f06e793160f0b5ec697df8641ff50b2